chore: move opstate
